requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
pdfreader==0.1.12
Pillow==9.5.0
pycryptodome==3.17
python-dateutil==2.8.2
python-docx==0.8.11
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2